optbench: move code under misc/python
- Move the `mzbench_opt` python package from
  `src/transform/benches/mzbench_opt` to
  `misc/python/materialize/optbench`.
- Move the `mzbench-opt` CLI from
  `src/transform/benches/mzbench_opt/cli.py` to
  `misc/python/materialize/cli/optbench.py`.
- Remove the setuptools-based `mzbench-opt` CLI
  in favor of a dedicated `bin/optbench` script.
