+++ bin/optbench
@@ -1,0 +1,14 @@
+#!/usr/bin/env bash
+
+# Copyright Materialize, Inc. and contributors. All rights reserved.
+#
+# Use of this software is governed by the Business Source License
+# included in the LICENSE file at the root of this repository.
+#
+# As of the Change Date specified in that file, in accordance with
+# the Business Source License, use of this software will be governed
+# by the Apache License, Version 2.0.
+#
+# dashboard-clean - Clean the dashboard model json to be fit for public consumption
+
+exec "$(dirname "$0")"/pyactivate --dev -m materialize.cli.optbench "$@"
