+++ ci/builder/requirements-dev.txt
@@ -5,16 +5,20 @@ flake8==4.0.1
 humanize==3.11.0
 isort==5.9.3
 mypy==0.910
+numpy==1.21.4
+pandas==1.3.4
 pdoc3==0.10.0
 prettytable==2.2.1
 psutil==5.8.0
 psycopg2==2.9.1
-types-psycopg2==2.9.1
 pytest==6.2.5
 requests==2.23.0
+sqlparse==0.4.2
 twine==3.4.2
+typer==0.4.0
 types-prettytable==2.1.0
 types-psutil==5.8.13
+types-psycopg2==2.9.1
 types-PyMYSQL==1.0.2
 types-PyYAML==5.4.10
 types-requests==2.25.11
