--- src/transform/benches/.gitignore
@@ -1,2 +1,0 @@
-*.egg-info
-build/
